# [Feature](mlu-ops): optimization for descritpor