--- a/core/tensor.cpp
+++ b/core/tensor.cpp
@@ -22,7 +22,7 @@
  *************************************************************************/
 #include <iomanip>
 #include <algorithm>
-
+#include <deque>
 #include "core/tensor.h"
 #include "core/logging.h"
 #include "core/type.h"
@@ -294,41 +294,62 @@ mluOpDestroySeqDataDescriptor(mluOpSeqDataDescriptor_t seq_data_desc) {
   return MLUOP_STATUS_SUCCESS;
 }
 
+namespace {
+
+#define MLUOP_TENSOR_QUEUE_ENABLE 1
+
 #if MLUOP_TENSOR_QUEUE_ENABLE
-static mluOpTensorDescriptorQueueStruct *queue_array = NULL;
-static std::hash<std::thread::id> hasher;
-
-MLUOP_ATTRIBUTE_CONSTRUCTOR MLUOP_ATTRIBUTE_VISIBILITY_HIDDEN void mluOpInit() {
-  if (!queue_array) {
-    queue_array =
-        new (std::nothrow) mluOpTensorDescriptorQueueStruct[QUEUE_ARRAY_LENGTH];
+struct mluOpTensorDescriptorQueueStruct {
+  mluOpTensorDescriptorQueueStruct() {
+    extend(extend_num);
+    extend_num *= 2;
   }
-}
-
-MLUOP_ATTRIBUTE_DESTRUCTOR MLUOP_ATTRIBUTE_VISIBILITY_HIDDEN void mluOpExit() {
-  if (queue_array) {
-    delete[] queue_array;
-    queue_array = NULL;
+  explicit mluOpTensorDescriptorQueueStruct(size_t n) {
+    extend_num = n;
+    extend(extend_num);
+    extend_num *= 2;
   }
-}
+
+  // Let the OS do the cleanup since it's a global variable
+  ~mluOpTensorDescriptorQueueStruct() {}
+
+  inline void lock() {
+    while (flag.test_and_set(std::memory_order_acquire)) {
+    }
+  }
+  inline void unlock() { flag.clear(std::memory_order_release); }
+  inline void extend(size_t n) {
+    mluOpTensorStruct *header = new (std::nothrow) mluOpTensorStruct[n];
+    for (size_t i = 0; i < n; ++i) {
+      mluOpTensorStruct *desc = header + i;
+      queue.push_front(desc);
+    }
+  }
+  size_t extend_num = 128;
+  std::deque<mluOpTensorDescriptor_t> queue;
+  std::atomic_flag flag = ATOMIC_FLAG_INIT;
+};
+
+static mluOpTensorDescriptorQueueStruct queue_array;
 #endif
+}  // anonymous namespace
+
 /* MLUOP interface */
 mluOpStatus_t MLUOP_WIN_API
 mluOpCreateTensorDescriptor(mluOpTensorDescriptor_t *desc) {
   PARAM_CHECK("[mluOpCreateTensorDescriptor]", desc != NULL);
 
 #if MLUOP_TENSOR_QUEUE_ENABLE
-  size_t id = hasher(std::this_thread::get_id()) % QUEUE_ARRAY_LENGTH;
-  queue_array[id].lock();
-  if (MLUOP_PREDICT_FALSE(queue_array[id].queue.empty())) {
-    queue_array[id].extend(queue_array[id].extend_num);
-    queue_array[id].extend_num *= 2;
+  queue_array.lock();
+  if MLUOP_PREDICT_FALSE (queue_array.queue.empty()) {
+    queue_array.extend(queue_array.extend_num);
+    queue_array.extend_num *= 2;
   }
-  *desc = queue_array[id].queue.front();
-  queue_array[id].queue.pop();
-  queue_array[id].unlock();
+  *desc = ::new (queue_array.queue.front()) mluOpTensorStruct;
+  queue_array.queue.pop_front();
+  queue_array.unlock();
 #else
-  mluOpTensorStruct *ts = new (std::nothrow) mluOpTensorStruct();
+  mluOpTensorStruct *ts = new (std::nothrow) mluOpTensorStruct;
   *desc = ts;
 #endif
 
@@ -341,32 +362,29 @@ mluOpStatus_t MLUOP_WIN_API mluOpCreateGroupTensorDescriptors(
   PARAM_CHECK("[mluOpCreateGroupTensorDescriptors]", desc_num > 0);
 
 #if MLUOP_TENSOR_QUEUE_ENABLE
-  size_t id = hasher(std::this_thread::get_id()) % QUEUE_ARRAY_LENGTH;
-  queue_array[id].lock();
-  if (MLUOP_PREDICT_FALSE(queue_array[id].queue.empty() ||
-                          (size_t)desc_num >
-                              (size_t)queue_array[id].queue.size())) {
-    queue_array[id].extend(
-        std::max((size_t)queue_array[id].extend_num, (size_t)desc_num));
-    queue_array[id].extend_num =
-        2 * std::max((size_t)queue_array[id].extend_num, (size_t)desc_num);
+  queue_array.lock();
+  if MLUOP_PREDICT_FALSE (queue_array.queue.size() < desc_num) {
+    queue_array.extend(std::max(queue_array.extend_num, (size_t)desc_num));
+    queue_array.extend_num =
+        2 * std::max(queue_array.extend_num, (size_t)desc_num);
   }
   for (int i = 0; i < desc_num; ++i) {
-    *(group_desc[i]) = queue_array[id].queue.front();
-    queue_array[id].queue.pop();
+    *(group_desc[i]) = queue_array.queue.front();
+    queue_array.queue.pop_front();
   }
-  queue_array[id].unlock();
+  queue_array.unlock();
 #else
   for (int i = 0; i < desc_num; ++i) {
-    mluOpTensorStruct *ts = new (std::nothrow) mluOpTensorStruct();
-    *(group_desc[i]) = ts;
+    mluOpTensorStruct *ts = new (std::nothrow) mluOpTensorStruct;
+    group_desc[i][0] = ts;
   }
 #endif
 
   return MLUOP_STATUS_SUCCESS;
 }
 
-mluOpStatus_t mluOpSetTensorDescriptorZeroDim(mluOpTensorDescriptor_t desc) {
+static inline mluOpStatus_t mluOpSetTensorDescriptorZeroDim(
+    mluOpTensorDescriptor_t desc) {
   if (desc->pointer_mode == MLUOP_POINTER_MODE_HOST) {
     desc->dim = 0;
     desc->total_element_num = 1;
@@ -422,32 +440,23 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptor_v2(
   }
 }
 
-mluOpStatus_t mluOpSetTensorDescriptorDimBase(mluOpTensorDescriptor_t desc,
-                                              int dimNb, const void *dimSize) {
-  PARAM_CHECK("[mluOpSetTensorDescriptorDim]", desc != NULL);
-  PARAM_CHECK("[mluOpSetTensorDescriptorDim]", dimNb > 0);
-  PARAM_CHECK("[mluOpSetTensorDescriptorDim]", dimSize != NULL);
-
-  desc->dim = dimNb;
-
-  if (MLUOP_PREDICT_FALSE(desc->larger_dims != NULL)) {
-    delete[] desc->larger_dims;
-    desc->larger_dims = NULL;
+// Internal interface. Caller should guarantee parameter validity.
+static inline void mluOpSetTensorDescriptorDimBase(mluOpTensorDescriptor_t desc,
+                                                   int dimNb) {
+  if (dimNb != desc->dim) {
+    if MLUOP_PREDICT_FALSE (desc->dims != desc->normal_dims) {
+      delete[] desc->dims;
+      delete[] desc->strides;
+    }
+    if MLUOP_PREDICT_FALSE (dimNb > MLUOP_DIM_MAX) {
+      desc->dims = new (std::nothrow) int64_t[dimNb];
+      desc->strides = new (std::nothrow) int64_t[dimNb];
+    } else {
+      desc->dims = desc->normal_dims;
+      desc->strides = desc->normal_strides;
+    }
+    desc->dim = dimNb;
   }
-  if (MLUOP_PREDICT_FALSE(desc->larger_strides != NULL)) {
-    delete[] desc->larger_strides;
-    desc->larger_strides = NULL;
-  }
-  if (MLUOP_PREDICT_FALSE(dimNb > MLUOP_DIM_MAX)) {
-    desc->larger_dims = new (std::nothrow) int64_t[dimNb];
-    desc->larger_strides = new (std::nothrow) int64_t[dimNb];
-    desc->dims = desc->larger_dims;
-    desc->strides = desc->larger_strides;
-  } else {
-    desc->dims = desc->normal_dims;
-    desc->strides = desc->normal_strides;
-  }
-  return MLUOP_STATUS_SUCCESS;
 }
 
 mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorDim(
@@ -456,8 +465,7 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorDim(
     CHECK_RETURN("[mluOpSetTensorDescriptorDim]",
                  mluOpSetTensorDescriptorZeroDim(desc));
   } else {
-    CHECK_RETURN("[mluOpSetTensorDescriptorDim]",
-                 mluOpSetTensorDescriptorDimBase(desc, dimNb, (void *)dimSize));
+    mluOpSetTensorDescriptorDimBase(desc, dimNb);
     std::copy(dimSize, dimSize + dimNb, desc->dims);
   }
 
@@ -493,8 +501,7 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorDim(
 
 mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorDim_v2(
     mluOpTensorDescriptor_t desc, int dimNb, const int64_t *dimSize) {
-  CHECK_RETURN("[mluOpSetTensorDescriptorDim]",
-               mluOpSetTensorDescriptorDimBase(desc, dimNb, (void *)dimSize));
+  mluOpSetTensorDescriptorDimBase(desc, dimNb);
 
   memcpy(desc->dims, dimSize, dimNb * sizeof(int64_t));
 
@@ -541,34 +548,30 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetGroupTensorDescriptors(
 
   int group_dimSize_iterator = 0;
   for (int i = 0; i < desc_num; ++i) {
-    (*(group_desc[i]))->dim = group_dimNb[i];
-    (*(group_desc[i]))->dtype = group_dtype[i];
-    (*(group_desc[i]))->layout = group_layout[i];
+    group_desc[i][0]->dim = group_dimNb[i];
+    group_desc[i][0]->dtype = group_dtype[i];
+    group_desc[i][0]->layout = group_layout[i];
 
     if (MLUOP_PREDICT_FALSE(group_dimNb[i] > MLUOP_DIM_MAX)) {
-      (*(group_desc[i]))->larger_dims =
-          new (std::nothrow) int64_t[group_dimNb[i]];
-      (*(group_desc[i]))->larger_strides =
-          new (std::nothrow) int64_t[group_dimNb[i]];
-      (*(group_desc[i]))->dims = (*(group_desc[i]))->larger_dims;
-      (*(group_desc[i]))->strides = (*(group_desc[i]))->larger_strides;
+      group_desc[i][0]->dims = new (std::nothrow) int64_t[group_dimNb[i]];
+      group_desc[i][0]->strides = new (std::nothrow) int64_t[group_dimNb[i]];
     } else {
-      (*(group_desc[i]))->dims = (*(group_desc[i]))->normal_dims;
-      (*(group_desc[i]))->strides = (*(group_desc[i]))->normal_strides;
+      group_desc[i][0]->dims = group_desc[i][0]->normal_dims;
+      group_desc[i][0]->strides = group_desc[i][0]->normal_strides;
     }
     std::copy(group_dimSize + group_dimSize_iterator,
               group_dimSize + group_dimSize_iterator + group_dimNb[i],
-              (*(group_desc[i]))->dims);
+              group_desc[i][0]->dims);
 
     // infer strides of dimNb dimensions and compute total_num and total_size
     int strideBase = 1;
     for (int j = group_dimNb[i] - 1; j >= 0; --j) {
-      (*(group_desc[i]))->strides[j] = strideBase;
-      strideBase *= (*(group_desc[i]))->dims[j];
+      group_desc[i][0]->strides[j] = strideBase;
+      strideBase *= group_desc[i][0]->dims[j];
     }
-    (*(group_desc[i]))->total_element_num = strideBase;
-    (*(group_desc[i]))->total_tensor_size =
-        (*(group_desc[i]))->total_element_num *
+    group_desc[i][0]->total_element_num = strideBase;
+    group_desc[i][0]->total_tensor_size =
+        group_desc[i][0]->total_element_num *
         mluop::getSizeOfDataType(group_dtype[i]);
 
     // compute new iterator for next loop.
@@ -591,33 +594,29 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetGroupTensorDescriptors_v2(
 
   int group_dimSize_iterator = 0;
   for (int i = 0; i < desc_num; ++i) {
-    (*(group_desc[i]))->dim = group_dimNb[i];
-    (*(group_desc[i]))->dtype = group_dtype[i];
-    (*(group_desc[i]))->layout = group_layout[i];
+    group_desc[i][0]->dim = group_dimNb[i];
+    group_desc[i][0]->dtype = group_dtype[i];
+    group_desc[i][0]->layout = group_layout[i];
 
     if (MLUOP_PREDICT_FALSE(group_dimNb[i] > MLUOP_DIM_MAX)) {
-      (*(group_desc[i]))->larger_dims =
-          new (std::nothrow) int64_t[group_dimNb[i]];
-      (*(group_desc[i]))->larger_strides =
-          new (std::nothrow) int64_t[group_dimNb[i]];
-      (*(group_desc[i]))->dims = (*(group_desc[i]))->larger_dims;
-      (*(group_desc[i]))->strides = (*(group_desc[i]))->larger_strides;
+      group_desc[i][0]->dims = new (std::nothrow) int64_t[group_dimNb[i]];
+      group_desc[i][0]->strides = new (std::nothrow) int64_t[group_dimNb[i]];
     } else {
-      (*(group_desc[i]))->dims = (*(group_desc[i]))->normal_dims;
-      (*(group_desc[i]))->strides = (*(group_desc[i]))->normal_strides;
+      group_desc[i][0]->dims = group_desc[i][0]->normal_dims;
+      group_desc[i][0]->strides = group_desc[i][0]->normal_strides;
     }
-    memcpy((*(group_desc[i]))->dims, group_dimSize + group_dimSize_iterator,
+    memcpy(group_desc[i][0]->dims, group_dimSize + group_dimSize_iterator,
            group_dimNb[i] * sizeof(int64_t));
 
     // infer strides of dimNb dimensions and compute total_num and total_size
     int strideBase = 1;
     for (int j = group_dimNb[i] - 1; j >= 0; --j) {
-      (*(group_desc[i]))->strides[j] = strideBase;
-      strideBase *= (*(group_desc[i]))->dims[j];
+      group_desc[i][0]->strides[j] = strideBase;
+      strideBase *= group_desc[i][0]->dims[j];
     }
-    (*(group_desc[i]))->total_element_num = strideBase;
-    (*(group_desc[i]))->total_tensor_size =
-        (*(group_desc[i]))->total_element_num *
+    group_desc[i][0]->total_element_num = strideBase;
+    group_desc[i][0]->total_tensor_size =
+        group_desc[i][0]->total_element_num *
         mluop::getSizeOfDataType(group_dtype[i]);
 
     // compute new iterator for next loop.
@@ -630,7 +629,28 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetGroupTensorDescriptors_v2(
 mluOpStatus_t MLUOP_WIN_API
 mluOpResetTensorDescriptor(mluOpTensorDescriptor_t desc) {
   PARAM_CHECK("[mluOpResetTensorDescriptor]", desc != NULL);
-  desc->reset();
+
+  if MLUOP_PREDICT_FALSE (desc->dims != desc->normal_dims) {
+    delete[] desc->dims;
+    desc->dims = desc->normal_dims;
+  }
+  if MLUOP_PREDICT_FALSE (desc->strides != desc->normal_strides) {
+    delete[] desc->strides;
+    desc->strides = desc->normal_strides;
+  }
+
+  desc->dim = 0;
+  desc->dtype = MLUOP_DTYPE_FLOAT;
+  desc->onchip_dtype = MLUOP_DTYPE_INVALID;
+  desc->layout = MLUOP_LAYOUT_ARRAY;
+  desc->pointer_mode = MLUOP_POINTER_MODE_DEVICE;
+
+  desc->total_element_num = 0;
+  desc->total_tensor_size = 0;
+
+  desc->position = 0;
+  desc->scale = 1.0f;
+  desc->offset = 0;
 
   return MLUOP_STATUS_SUCCESS;
 }
@@ -653,7 +673,7 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorEx(
     PARAM_CHECK("[mluOpSetTensorDescriptorEx]", dimStride != NULL);
     PARAM_CHECK("[mluOpSetTensorDescriptorEx]", dimNb > 0);
 
-    mluOpSetTensorDescriptorDimBase(desc, dimNb, (void *)dimSize);
+    mluOpSetTensorDescriptorDimBase(desc, dimNb);
     std::copy(dimSize, dimSize + dimNb, desc->dims);
     std::copy(dimStride, dimStride + dimNb, desc->strides);
 
@@ -680,14 +700,13 @@ mluOpStatus_t MLUOP_WIN_API mluOpSetTensorDescriptorEx_v2(
   desc->dtype = dtype;
   desc->layout = layout;
 
-  if (dimNb == 0) {
+  if MLUOP_PREDICT_FALSE (dimNb == 0) {
     return mluOpSetTensorDescriptorZeroDim(desc);
   } else {
     PARAM_CHECK("[mluOpSetTensorDescriptorEx]", dimSize != NULL);
     PARAM_CHECK("[mluOpSetTensorDescriptorEx]", dimStride != NULL);
-    PARAM_CHECK("[mluOpSetTensorDescriptorEx]", dimNb > 0);
 
-    mluOpSetTensorDescriptorDimBase(desc, dimNb, (void *)dimSize);
+    mluOpSetTensorDescriptorDimBase(desc, dimNb);
     memcpy(desc->dims, dimSize, dimNb * sizeof(int64_t));
     memcpy(desc->strides, dimStride, dimNb * sizeof(int64_t));
 
@@ -872,13 +891,12 @@ mluOpStatus_t MLUOP_WIN_API mluOpGetTensorDescriptorPointerMode(
 mluOpStatus_t MLUOP_WIN_API
 mluOpDestroyTensorDescriptor(mluOpTensorDescriptor_t desc) {
   PARAM_CHECK("[mluOpDestroyTensorDescriptor]", desc != NULL);
-  desc->reset();
 
 #if MLUOP_TENSOR_QUEUE_ENABLE
-  size_t id = hasher(std::this_thread::get_id()) % QUEUE_ARRAY_LENGTH;
-  queue_array[id].lock();
-  queue_array[id].queue.emplace(desc);
-  queue_array[id].unlock();
+  queue_array.lock();
+  desc->~mluOpTensorStruct();
+  queue_array.queue.push_front(desc);
+  queue_array.unlock();
 #else
   delete desc;
 #endif
@@ -892,17 +910,15 @@ mluOpStatus_t MLUOP_WIN_API mluOpDestroyGroupTensorDescriptors(
   PARAM_CHECK("[mluOpDestroyGroupTensorDescriptors]", desc_num > 0);
 
 #if MLUOP_TENSOR_QUEUE_ENABLE
-  size_t id = hasher(std::this_thread::get_id()) % QUEUE_ARRAY_LENGTH;
-  queue_array[id].lock();
+  queue_array.lock();
   for (int i = 0; i < desc_num; ++i) {
-    (*(group_desc[i]))->reset();
-    queue_array[id].queue.emplace(*(group_desc[i]));
+    group_desc[i][0]->~mluOpTensorStruct();
+    queue_array.queue.push_front(group_desc[i][0]);
   }
-  queue_array[id].unlock();
+  queue_array.unlock();
 #else
   for (int i = 0; i < desc_num; ++i) {
-    (*(group_desc[i]))->reset();
-    delete (*(group_desc[i]));
+    delete group_desc[i][0];
   }
 #endif
 
@@ -913,9 +929,7 @@ mluOpStatus_t MLUOP_WIN_API mluOpDestroyGroupTensorDescriptors(
 uint64_t MLUOP_WIN_API
 mluOpGetTensorElementNum(const mluOpTensorDescriptor_t desc) {
   CHECK(desc != NULL);
-  uint64_t tensor_num = 1;
-  auto return_status = desc->tensorElementsNumber(tensor_num);
-  return tensor_num;
+  return desc->total_element_num;
 }
 
 uint64_t mluOpGetSeqDataElementNum(mluOpSeqDataDescriptor_t desc) {