--- a/core/type.h
+++ b/core/type.h
@@ -52,9 +52,38 @@ static mluOpStatus_t getLowAndHighValueFrom64Bits(T value, uint32_t* high,
   return MLUOP_STATUS_SUCCESS;
 }
 
-// TODO(None) use const char * instead of std::string (need mluop_extra
-// fix first) hide visibility
-size_t MLUOP_WIN_API getSizeOfDataType(mluOpDataType_t dtype);
+static inline size_t MLUOP_WIN_API getSizeOfDataType(mluOpDataType_t dtype) {
+  switch (dtype) {
+    default: {
+      return 0;
+    }
+    case MLUOP_DTYPE_BOOL:
+    case MLUOP_DTYPE_INT8:
+    case MLUOP_DTYPE_UINT8: {
+      return 1;
+    }
+    case MLUOP_DTYPE_INT16:
+    case MLUOP_DTYPE_UINT16:
+    case MLUOP_DTYPE_HALF:
+    case MLUOP_DTYPE_BFLOAT16: {
+      return 2;
+    }
+    case MLUOP_DTYPE_INT31:
+    case MLUOP_DTYPE_INT32:
+    case MLUOP_DTYPE_UINT32:
+    case MLUOP_DTYPE_FLOAT:
+    case MLUOP_DTYPE_COMPLEX_HALF: {
+      return 4;
+    }
+    case MLUOP_DTYPE_UINT64:
+    case MLUOP_DTYPE_INT64:
+    case MLUOP_DTYPE_DOUBLE:
+    case MLUOP_DTYPE_COMPLEX_FLOAT: {
+      return 8;
+    }
+  }
+}
+
 std::string MLUOP_WIN_API getNameOfDataType(mluOpDataType_t dtype);  // NOLINT
 std::string MLUOP_WIN_API getNameOfTensorLayout(mluOpTensorLayout_t layout);  // NOLINT