@nth-BYTE 2024-09-23
jira:http://jira.cambricon.com/browse/CNNLCORE-18421